Builtins: add exit, abort, @program-count, @variable-count
The @-builtin @program-count aids in figuring out if a call is
tail recursive or not. The variable count can also be used to
see if tail recursion is active.

If there is active TCO then these values do not grow unbounded.

An idea I've been munching on is to let all builtins return
an Option<String>. The string contains an error message. Eval
can then use the return value and construct an error and unwind.

This would make it possible to add line, column, and source to the
message consistently.