Change Transfer by appending -> Option<String>
A return of None denotes that execution went according to plan,
but a return of Some notifies the VM of an error that occurred.

The idea is to let all builtins issue their own custom error
messages, but of course we'll have a convention for it.

So for example fn not would say: Some("not: arity, expected 1, got 0").
The VM would append a callstack to the error. Honestly I never understood
why that wasn't a standard in so many other languages. Just show me the
damn call stack and I can much more easily figure out where the bug is.

Now of course TCO comes in the way and demolishes some of the call stack
but we should be fine in most cases, it will give a general idea of
control flow.

I think a cool format for the error messages is to write:

	<custom error here> <= source1 <= source2 <= ... <= sourcen

So we first see the error, then the closest call site to the error, then
the second closest, and so on. I'll have to think about it.