Formalize - to some degree - the errors
The errors generated by builtins and other inconsistencies
are now reported to err. Err appends the current
source to that error and unwinds.

I like this design, but I wish it appended the stack trace.
The nice thing about a stack trace is that it's so useful...
So that's what's left, plus some other stuff.

There's also an issue with performane, I benchmarked the program
and it was really bad compared to other languages.
I mean, 100'000'000 iterations in 5 minutes makes for about
330k iterations per second. You know, still within reason
for a glue language I suppose but it's much slower than python.
But I think optimizations have a lot to say, as well as processor
architecture. Look, the corei7 that I run it on finishes within 50
seconds which is about 6x increase in speed compared to another
processor. So I'll need to look into that.

BUT we all know premature optimization is the root of all evil.
So to optimize right now would kill new ideas. I'll just keep
going at the current pace and see where it leads me.

Which marks my last point that it's virtually impossible to create
a compiler for this language. Seriously think about it, every single
variable can be redefined to something else and you can't know when
or where this happens, and so what do you do?
DON'T COMPILE is the obvious answer because we CAN'T. Although
I think that if we use some really clever tools we can prove
certain code paths and thus certain properties.

These properties are essential in determining the information
necessary for compilation. So you know,... I think there's
a slight chance for compilation to be possible... but...
well let's just do that another time.