Implement the write functionality for arbitrary data
The idea behind write is that one can do (write (read))
and get the exact same representation of data. That is;

	(write (quote a b c))

will write "(a b c)" to the output (without ") and read
will construct a list of symbols containing a, b, and c.

What read does not do is evaluate, however, we say that
some forms are self-evaluating, like numbers and various
functions:

	(write (fn () 0))

writes "(fn () 0)", read will read it in as a list
of symbols fn, (), and 0. By stating that this form itself
is  self-evaluating, we can safely read-write this as much
as we want.

Furthermore I've taken the liberty to remove Userdata..
for now. :]