Formalize the stack trace as a list of lists
Tracing isn't supposed to be done very often so the overhead of
consing is negligible. This also makes the traces actually workable,
for example teko-rs can colorize certain parts of the trace quite
easily now.

Of course it looks horrible at the moment, but that's why we should
implement a pretty-printer that removes the quotes and much other junk.

The source string has to be a string since a source is an arbitrary string,
and so it can't be a symbol. I am thinking though that it might be possible
to just have the code be in the trace, since it consists of symbols only...