Replace Rc::new with rc and clean up fn err
Right now err will create the stack trace as an actual in-memory
list object. There is no mixing of strings. The only "string" is the
error message that comes in, but that's by design and unavoidable,... well
I guess we COULD return arbitrary objects in the error trace but, is that
really useful?