Fix union operation
The union operation is performed during each tail call
optimization. The previous version would not work correctly
in all cases, it was a logical error. This is fixed now.

When a is [] and b is ["something"], something would never
be part of the union since we iterate over a. Now we iterate
over b instead.