Remove all user-based data
User data was an initial idea that allows users to extend the implementation with their own data types.

The problem with this is that we then need to have the users implement a lot of details like serialization. Instead, if we want to handle data from other programs we can use strings and handles.

Not sure what problems require more data than a table, vector, and list. Even so, new data structures can be added in the future, but it has to be a language standard. The language is supposed to be nearly perfect, and should not be polluted by unnecessary details.