Update goals
We do want multithreading with channels. This is very useful, as it allows the program to communicate with the rest of the system.

The idea here is to not have any mutable sharing, but instead use messages. These are Arc<Data>, so they can safely be shared and it's essentially just an Rc pointer, so it's extremely cheap even to share huge data structures.

For channels the idea is to let each thread return a handle (integer) that specifies which thread it is. Optionally, we can create a builtin data type of thread, hell, we can even encode the entire function into it. That'd be cool actually. Anyway, we need facilities for stopping, pausing, and resuming a thread, and a single "mailbox".

I'd like multiple channels actually, but having a single communication channel seems easier:

	(local other
		(thread
		  (function ()
		    (local x (receive))
		    (write x))))

	(send data other)

Additionally, custom channels can be made. But, I'd like stuff to be serializable. So that means that a thread will need to be loadable. One way of doing this is to also specify a unique identifier. This way, we can load threads from files, and if the function and id match, we know that we can check for such a thread.